refine storage code organization
