+++ arch/sec_hw/mailbox_interface/mailbox_storage.c
@@ -14,83 +14,10 @@
 #include "arch/ring_buffer.h"
 #include "arch/octopos_mbox_owner_map.h"
 #include "arch/octopos_mbox.h"
-#include "arch/syscall.h"
 
 #include "octopos/mailbox.h"
 #include "octopos/storage.h"
 #include "octopos/error.h"
-
-#define HW_MAILBOX_BLOCKING
-
-
-#define STORAGE_SET_ONE_RET(ret0)	\
-	SERIALIZE_32(ret0, &buf[0])
-
-#define STORAGE_SET_TWO_RETS(ret0, ret1)	\
-	SERIALIZE_32(ret0, &buf[0])				\
-	SERIALIZE_32(ret1, &buf[4])
-
-/* FIXME: when calling this one, we need to allocate a ret_buf. Can we avoid that? */
-#define STORAGE_SET_ONE_RET_DATA(ret0, data, size)		\
-	*((uint32_t *) &buf[0]) = ret0;				\
-	uint8_t max_size = MAILBOX_QUEUE_MSG_SIZE - 5;		\
-	if (max_size < 256 && size <= ((int) max_size)) {	\
-		buf[4] = (uint8_t) size;			\
-		memcpy(&buf[5], data, size);			\
-	} else {						\
-		printf("Error: invalid max_size or size\n");	\
-		buf[4] = 0;					\
-	}
-
-#define STORAGE_GET_ONE_ARG		\
-	uint32_t arg0;			\
-	arg0 = *((uint32_t *) &buf[1]); \
-
-#define STORAGE_GET_TWO_ARGS		\
-	uint32_t arg0, arg1;		\
-	arg0 = *((uint32_t *) &buf[1]); \
-	arg1 = *((uint32_t *) &buf[5]); \
-
-#define STORAGE_GET_THREE_ARGS		\
-	uint32_t arg0, arg1, arg2;	\
-	arg0 = *((uint32_t *) &buf[1]); \
-	arg1 = *((uint32_t *) &buf[5]); \
-	arg2 = *((uint32_t *) &buf[9]);\
-
-#define STORAGE_GET_ZERO_ARGS_DATA				\
-	uint8_t data_size, *data;				\
-	uint8_t max_size = MAILBOX_QUEUE_MSG_SIZE - 2;		\
-	if (max_size >= 256) {					\
-		printf("Error: max_size not supported\n");	\
-		STORAGE_SET_ONE_RET((uint32_t) ERR_INVALID)	\
-		return;						\
-	}							\
-	data_size = buf[1];					\
-	if (data_size > max_size) {				\
-		printf("Error: size not supported\n");		\
-		STORAGE_SET_ONE_RET((uint32_t) ERR_INVALID)	\
-		return;						\
-	}							\
-	data = &buf[2];					\
-
-#define STORAGE_GET_ONE_ARG_DATA				\
-	uint32_t arg0;						\
-	uint8_t data_size, *data;				\
-	arg0 = *((uint32_t *) &buf[1]);				\
-	uint8_t max_size = MAILBOX_QUEUE_MSG_SIZE - 6;		\
-	if (max_size >= 256) {					\
-		printf("Error: max_size not supported\n");	\
-		STORAGE_SET_ONE_RET((uint32_t) ERR_INVALID)	\
-		return;						\
-	}							\
-	data_size = buf[5];					\
-	if (data_size > max_size) {				\
-		printf("Error: size not supported\n");		\
-		STORAGE_SET_ONE_RET((uint32_t) ERR_INVALID)	\
-		return;						\
-	}							\
-	data = &buf[6];
-
 
 XIntc			intc;
 
@@ -107,32 +34,17 @@ XMbox*			Mbox_regs[NUM_QUEUES + 1];
 UINTPTR			Mbox_ctrl_regs[NUM_QUEUES + 1];
 
 static FATFS	fatfs;
-
-
-static uint8_t 	*sketch_buffer[NUM_QUEUES + 1];
-static u32 		sketch_buffer_offset[NUM_QUEUES + 1];
-
-bool 			is_queue_set_bound = false;
-int 			bound_partition = -1;
-
-uint8_t			config_key[STORAGE_KEY_SIZE];
-bool			is_config_locked = false;
+BYTE			work[FF_MAX_SS];
 
 u32				DEBUG_STATUS_REGISTERS[30] = {0};
 
-extern struct	partition partitions[NUM_PARTITIONS];
-extern uint32_t	partition_sizes[NUM_PARTITIONS]; 
+void process_request(uint8_t *buf);
+void initialize_storage_space(void);
 
-_Bool handle_partial_message(uint8_t *message_buffer, uint8_t *queue_id, u32 bytes_read);
-
-/* https://stackoverflow.com/questions/7775027/how-to-create-file-of-x-size */
-static void initialize_storage_space(void)
+static void initialize_ramfs(void)
 {
 	TCHAR *Path = "0:/";
-	BYTE work[FF_MAX_SS];
 	FRESULT result;
-	FIL filep, filep2;
-	UINT NumBytesRead = 0, NumBytesWritten = 0;
 
 	result = f_mount(&fatfs, Path, 0);
 	if (result != FR_OK) {
@@ -145,567 +57,16 @@ static void initialize_storage_space(void)
 		SEC_HW_DEBUG_HANG();
 		return;
 	}
-
-	for (int i = 0; i < NUM_PARTITIONS; i++) {
-		struct partition *partition;
-		int suffix = i;
-		partition = &partitions[i];
-		partition->size = partition_sizes[i];
-
-		memset(partition->data_name, 0x0, 256);
-		sprintf(partition->data_name, "octopos_partition_%d_data", suffix);
-
-		memset(partition->create_name, 0x0, 256);
-		sprintf(partition->create_name, "octopos_partition_%d_create", suffix);
-
-		memset(partition->lock_name, 0x0, 256);
-		sprintf(partition->lock_name, "octopos_partition_%d_lock", suffix);
-
-		result = f_open(&filep, partition->data_name, FA_READ);
-		if (result) {
-			/* create empty file */
-			f_open(&filep2, partition->data_name, FA_CREATE_ALWAYS | FA_WRITE);
-			/* populate with zeros (so that first read doesn't return an error */
-			uint8_t zero_block[STORAGE_BLOCK_SIZE];
-			memset(zero_block, 0x0, STORAGE_BLOCK_SIZE);
-			f_lseek(&filep2, 0);
-			for (uint32_t j = 0; j < partition->size; j++)
-				f_write(&filep2, (const void*) zero_block, STORAGE_BLOCK_SIZE, &NumBytesWritten);
-
-			f_close(&filep2);
-		}
-		f_close(&filep);
-		/* Is partition created? */
-		result = f_open(&filep, partition->create_name, FA_READ);
-		if (result) {
-			/* create empty file */
-			f_open(&filep2, partition->create_name, FA_CREATE_ALWAYS | FA_WRITE);
-			f_close(&filep2);
-			/* Also wipe lock info (which should not have any valid key anyway. */
-			f_open(&filep2, partition->lock_name, FA_CREATE_ALWAYS | FA_WRITE);
-			f_close(&filep2);
-			partition->is_locked = false;
-			continue;
-		}
-		f_lseek(&filep, 0);
-		uint32_t tag = 0;
-		f_read(&filep, &tag, 4, &NumBytesRead);
-		f_close(&filep);
-		if (NumBytesRead == 4 && tag == 1) {
-			partition->is_created = true;
-		} else {
-			/* create empty file */
-			f_open(&filep2, partition->create_name, FA_CREATE_ALWAYS | FA_WRITE);
-			f_close(&filep2);
-			/* Also wipe any key info. This should not normally happen. */
-			f_open(&filep2, partition->lock_name, FA_CREATE_ALWAYS | FA_WRITE);
-			f_close(&filep2);
-			partition->is_locked = false;
-			continue;
-		}
-		/* lock partitions that have an active key */
-		result = f_open(&filep, partition->lock_name, FA_READ);
-		if (result) {
-			/* create empty file */
-			f_open(&filep2, partition->lock_name, FA_CREATE_ALWAYS | FA_WRITE);
-			f_close(&filep2);
-			partition->is_locked = false;
-			continue;
-		}
-		uint8_t key[STORAGE_KEY_SIZE];
-		f_lseek(&filep, 0);
-		f_read(&filep, key, STORAGE_KEY_SIZE, &NumBytesRead);
-		f_close(&filep);
-		if (NumBytesRead == STORAGE_KEY_SIZE) {
-			partition->is_locked = true;
-		} else {
-			/* wipe lock file */
-			f_open(&filep2, partition->lock_name, FA_CREATE_ALWAYS | FA_WRITE);
-			f_close(&filep2);
-			partition->is_locked = false;
-		}
-	}
 }
 
-
-static int set_partition_key(uint8_t *data, int partition_id)
+void read_data_from_queue(uint8_t *buf, uint8_t queue_id)
 {
-	FRESULT result;
-	FIL filep;
-	UINT NumBytesWritten = 0;
-
-	result = f_open(&filep, partitions[partition_id].lock_name, FA_READ | FA_WRITE);
-	if (result) {
-		_SEC_HW_ERROR("%s: Error: couldn't open %s\n", __func__, partitions[partition_id].lock_name);
-		return ERR_FAULT;
-	}
-
-	f_lseek(&filep, 0);
-	f_write(&filep, (const void*)data, STORAGE_KEY_SIZE, &NumBytesWritten);
-	if (NumBytesWritten < STORAGE_KEY_SIZE) {
-		/* make sure to delete what was written */
-		f_open(&filep, partitions[partition_id].lock_name, FA_CREATE_ALWAYS | FA_WRITE);
-		f_close(&filep);
-		return ERR_FAULT;
-	}
-
-	f_close(&filep);
-	return 0;
+	XMbox_ReadBlocking(Mbox_regs[queue_id], (u32*) buf, MAILBOX_QUEUE_MSG_SIZE_LARGE);
 }
 
-static int remove_partition_key(int partition_id)
+void write_data_to_queue(uint8_t *buf, uint8_t queue_id)
 {
-	FIL filep;
-	FRESULT result;
-
-	result = f_open(&filep, partitions[partition_id].lock_name, FA_CREATE_ALWAYS | FA_WRITE);
-
-	if (result) {
-		_SEC_HW_ERROR("%s: Error: couldn't open %s\n", __func__, partitions[partition_id].lock_name);
-		return ERR_FAULT;
-	}
-	f_close(&filep);
-	return 0;
-}
-
-static int unlock_partition(uint8_t *data, int partition_id)
-{
-	FIL filep;
-	FRESULT result;
-	UINT NumBytesRead = 0;
-	uint8_t key[STORAGE_KEY_SIZE];
-
-	result = f_open(&filep, partitions[partition_id].lock_name, FA_READ);
-
-	if (result) {
-		_SEC_HW_ERROR("%s: Error: couldn't open %s\n", __func__, partitions[partition_id].lock_name);
-		return ERR_FAULT;
-	}
-
-	f_lseek(&filep, 0);
-	f_read(&filep, (void*)key, STORAGE_KEY_SIZE, &NumBytesRead);
-	f_close(&filep);
-	if (NumBytesRead != STORAGE_KEY_SIZE) {
-		/* TODO: if the key file is corrupted, then we might need to unlock, otherwise, we'll lose the partition. */
-		return ERR_FAULT;
-	}
-
-	for (int i = 0; i < STORAGE_KEY_SIZE; i++) {
-		if (key[i] != data[i]) {
-			return ERR_INVALID;
-		}
-	}
-
-	partitions[partition_id].is_locked = false;
-	return 0;
-}
-
-static int wipe_partition(int partition_id)
-{
-	FIL filep;
-	FRESULT result;
-	FILINFO finfo;
-	UINT NumBytesWritten = 0;
-	uint8_t zero_buf[STORAGE_BLOCK_SIZE] = {0};
-
-	f_stat(partitions[partition_id].data_name, &finfo);
-
-	result = f_open(&filep, partitions[partition_id].data_name, FA_CREATE_ALWAYS | FA_WRITE);
-
-	if (result) {
-		_SEC_HW_ERROR("%s: Error: couldn't open %s\n", __func__, partitions[partition_id].data_name);
-		return ERR_FAULT;
-	}
-	f_write(&filep, (const void*)zero_buf, finfo.fsize, &NumBytesWritten);
-	f_close(&filep);
-	return 0;
-}
-
-
-static void process_request(uint8_t *buf)
-{
-	FIL filep, filep2;
-	FRESULT result;
-	UINT NumBytesRead = 0, NumBytesWritten = 0;
-	uint8_t queue_id = 0;
-	u32 bytes_read;
-
-	/* write */
-	if (buf[0] == STORAGE_OP_WRITE) {
-		if (!is_queue_set_bound) {
-			_SEC_HW_ERROR("%s: Error: no partition is bound to queue set\n", __func__);
-			STORAGE_SET_ONE_RET(0)
-			return;
-		}
-		int partition_id = bound_partition;
-
-		if (partition_id < 0 || partition_id >= NUM_PARTITIONS) {
-			_SEC_HW_ERROR("%s: Error: invalid partition ID\n", __func__);
-			STORAGE_SET_ONE_RET(0)
-			return;
-		}
-		if (partitions[partition_id].is_locked) {
-			_SEC_HW_ERROR("%s: Error: partition is locked\n", __func__);
-			STORAGE_SET_ONE_RET(0)
-			return;
-		}
-		result = f_open(&filep, partitions[partition_id].data_name, FA_READ | FA_WRITE);
-		if (result) {
-			_SEC_HW_ERROR("%s: Error: couldn't open %s for write\n", __func__, 
-				partitions[partition_id].data_name);
-			STORAGE_SET_ONE_RET(0)
-			return;
-		}
-		STORAGE_GET_TWO_ARGS
-		uint32_t start_block = arg0;
-		uint32_t num_blocks = arg1;
-		if (start_block + num_blocks > partitions[partition_id].size) {
-			printf("%s: Error: invalid args\n", __func__);
-			STORAGE_SET_ONE_RET(0)
-			f_close(&filep);
-			return;
-		}
-		uint32_t seek_off = start_block * STORAGE_BLOCK_SIZE;
-		f_lseek(&filep, seek_off);
-
-		uint8_t data_buf[STORAGE_BLOCK_SIZE];
-		uint32_t size = 0;
-		for (uint32_t i = 0; i < num_blocks; i++) {
-#ifdef HW_MAILBOX_BLOCKING
-			XMbox_ReadBlocking(&Mbox_storage_data_in, (u32*) data_buf, MAILBOX_QUEUE_MSG_SIZE_LARGE);
-#else
-			queue_id = Q_STORAGE_DATA_IN;
-			XMbox_Read(&Mbox_storage_data_in,
-				(u32*) data_buf,
-				MAILBOX_QUEUE_MSG_SIZE_LARGE,
-				&bytes_read);
-			if (bytes_read != MAILBOX_QUEUE_MSG_SIZE_LARGE &&
-				!handle_partial_message(buf, &queue_id, bytes_read))
-#endif
-			f_write(&filep, (const void*)data_buf, STORAGE_BLOCK_SIZE, &NumBytesWritten);
-			size += (uint32_t) NumBytesWritten;
-		}
-		STORAGE_SET_ONE_RET(size);
-		f_close(&filep);
-	} else if (buf[0] == STORAGE_OP_READ) { /* read */
-		if (!is_queue_set_bound) {
-			_SEC_HW_ERROR("%s: Error: no partition is bound to queue set\n", __func__);
-			SEC_HW_DEBUG_HANG();
-			STORAGE_SET_ONE_RET(0)
-			return;
-		}
-		int partition_id = bound_partition;
-
-		if (partition_id < 0 || partition_id >= NUM_PARTITIONS) {
-			_SEC_HW_ERROR("%s: Error: invalid partition ID\n", __func__);
-			SEC_HW_DEBUG_HANG();
-			STORAGE_SET_ONE_RET(0)
-			return;
-		}
-
-		if (partitions[partition_id].is_locked) {
-			_SEC_HW_ERROR("%s: Error: partition is locked\n", __func__);
-			SEC_HW_DEBUG_HANG();
-			STORAGE_SET_ONE_RET(0)
-			return;
-		}
-
-		result = f_open(&filep, partitions[partition_id].data_name, FA_READ);
-		if (result) {
-			_SEC_HW_ERROR("%s: Error: couldn't open %s for read\n", __func__, partitions[partition_id].data_name);
-			SEC_HW_DEBUG_HANG();
-			STORAGE_SET_ONE_RET(0)
-			return;
-		}
-
-		STORAGE_GET_TWO_ARGS
-		uint32_t start_block = arg0;
-		uint32_t num_blocks = arg1;
-		if (start_block + num_blocks > partitions[partition_id].size) {
-			_SEC_HW_ERROR("%s: Error: invalid args\n", __func__);
-			SEC_HW_DEBUG_HANG();
-			STORAGE_SET_ONE_RET(0)
-			f_close(&filep);
-			return;
-		}
-
-		uint32_t seek_off = start_block * STORAGE_BLOCK_SIZE;
-		f_lseek(&filep, seek_off);
-		uint8_t data_buf[STORAGE_BLOCK_SIZE] = {0};
-		uint32_t size = 0;
-		for (uint32_t i = 0; i < num_blocks; i++) {
-			f_read(&filep, data_buf, STORAGE_BLOCK_SIZE, &NumBytesRead);
-			size += (uint32_t) NumBytesRead;
-			XMbox_WriteBlocking(&Mbox_storage_data_out, (u32*) data_buf, MAILBOX_QUEUE_MSG_SIZE_LARGE);
-		}
-
-		STORAGE_SET_ONE_RET(size);
-		f_close(&filep);
-	} else if (buf[0] == STORAGE_OP_SET_KEY) {
-		if (!is_queue_set_bound) {
-			_SEC_HW_ERROR("%s: Error: no partition is bound to queue set\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-
-		int partition_id = bound_partition;
-
-		if (partition_id < 0 || partition_id >= NUM_PARTITIONS) {
-			_SEC_HW_ERROR("%s: Error: invalid partition ID\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-
-		if (partitions[partition_id].is_locked) {
-			_SEC_HW_ERROR("%s: Error: can't set the key for a locked partition\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-
-		STORAGE_GET_ZERO_ARGS_DATA
-		if (data_size != STORAGE_KEY_SIZE) {
-			_SEC_HW_ERROR("%s: Error: incorrect key size\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-
-		uint32_t ret = (uint32_t) set_partition_key(data, partition_id);
-		STORAGE_SET_ONE_RET(ret)
-	} else if (buf[0] == STORAGE_OP_UNLOCK) {
-		STORAGE_GET_ZERO_ARGS_DATA
-
-		if (data_size != STORAGE_KEY_SIZE) {
-			_SEC_HW_ERROR("%s: Error: incorrect key size (sent for unlocking)\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-
-		int partition_id = -1;
-
-		for (int i = 0; i < NUM_PARTITIONS; i++) {
-			int ret = unlock_partition(data, i);
-			if (!ret) {
-				partition_id = i;
-				break;
-			}
-		}
-
-		if (partition_id < 0 || partition_id >= NUM_PARTITIONS) {
-			STORAGE_SET_ONE_RET(ERR_EXIST)
-			return;
-		}
-
-		if (partitions[partition_id].is_locked) {
-			STORAGE_SET_ONE_RET(ERR_FAULT)
-			return;
-		}
-
-		bound_partition = partition_id;
-		is_queue_set_bound = true;
-
-		STORAGE_SET_ONE_RET(0)
-	} else if (buf[0] == STORAGE_OP_LOCK) {
-		if (!is_queue_set_bound) {
-			_SEC_HW_ERROR("%s: Error: no partition is bound to queue set\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-
-		int partition_id = bound_partition;
-
-		if (partition_id < 0 || partition_id >= NUM_PARTITIONS) {
-			_SEC_HW_ERROR("%s: Error: invalid partition ID\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-
-		partitions[partition_id].is_locked = true;
-		bound_partition = -1;
-		is_queue_set_bound = false;
-
-		STORAGE_SET_ONE_RET(0)
-	} else if (buf[0] == STORAGE_OP_WIPE) {
-		if (!is_queue_set_bound) {
-			_SEC_HW_ERROR("%s: Error: no partition is bound to queue set\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-
-		int partition_id = bound_partition;
-
-		if (partition_id < 0 || partition_id >= NUM_PARTITIONS) {
-			_SEC_HW_ERROR("%s: Error: invalid partition ID\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-		if (partitions[partition_id].is_locked) {
-			_SEC_HW_ERROR("%s: Error: partition is locked\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-		int ret = wipe_partition(partition_id);
-		if (ret) {
-			_SEC_HW_ERROR("%s: Error: couldn't wipe the partition\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_FAULT)
-			return;
-		}
-
-		ret = remove_partition_key(partition_id);
-		if (ret) {
-			_SEC_HW_ERROR("%s: Error: couldn't remove partition key\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_FAULT)
-			return;
-		}
-
-		bound_partition = -1;
-		is_queue_set_bound = false;
-
-		STORAGE_SET_ONE_RET(0)
-	} else if (buf[0] == STORAGE_OP_CREATE_SECURE_PARTITION) {
-		if (is_config_locked) {
-			_SEC_HW_ERROR("%s: Error: config is locked (create partition op)\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-		STORAGE_GET_ONE_ARG_DATA
-		if (data_size != STORAGE_KEY_SIZE) {
-			_SEC_HW_ERROR("%s: Error: incorrect key size\n", __func__);
-			STORAGE_SET_TWO_RETS(ERR_INVALID, 0)
-			return;
-		}
-		uint32_t partition_size = arg0;
-
-		int partition_id = -1;
-
-		for (int i = 0; i < NUM_PARTITIONS; i++) {
-			if (!partitions[i].is_created && partitions[i].size == partition_size) {
-				partition_id = i;
-				break;
-			}
-		}
-
-		if (partition_id < 0 || partition_id >= NUM_PARTITIONS) {
-			_SEC_HW_ERROR("%s: Error: no partitions with the requested size available\n", __func__);
-			STORAGE_SET_TWO_RETS(ERR_AVAILABLE, 0)
-			return;
-		}
-
-		int ret = set_partition_key(data, partition_id);
-		if (ret) {
-			STORAGE_SET_TWO_RETS(ERR_FAULT, 0)
-			return;
-		}
-
-		result = f_open(&filep, partitions[partition_id].create_name, FA_READ | FA_WRITE);
-		if (result) {
-			STORAGE_SET_TWO_RETS(ERR_FAULT, 0)
-			return;
-		}
-		f_lseek(&filep, 0);
-		uint32_t tag = 1;
-		f_write(&filep, (const void*)&tag, 4, &NumBytesWritten);
-		uint32_t size = NumBytesWritten;
-		f_close(&filep);
-		if (size != 4) {
-			STORAGE_SET_TWO_RETS(ERR_FAULT, 0)
-			if (size > 0) { /* partial write */
-				/* wipe the file */
-				f_open(&filep2, partitions[partition_id].create_name, FA_CREATE_ALWAYS | FA_WRITE);
-				f_close(&filep2);
-			}
-			return;
-		}
-		partitions[partition_id].is_created = true;
-		STORAGE_SET_TWO_RETS(0, partition_id)
-	} else if (buf[0] == STORAGE_OP_DELETE_SECURE_PARTITION) {
-		if (is_config_locked) {
-			_SEC_HW_ERROR("%s: Error: config is locked (delete partition op)\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-
-		STORAGE_GET_ONE_ARG
-		uint32_t partition_id = arg0;
-
-		if (partition_id >= NUM_PARTITIONS) {
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-
-		if (!partitions[partition_id].is_created) {
-			_SEC_HW_ERROR("%s: Error: secure partition does not exist\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_EXIST)
-			return;
-		}
-
-		if (is_queue_set_bound && (bound_partition == (int) partition_id)) {
-			_SEC_HW_ERROR("%s: Error: secure partition currently bound to a queue set\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-		if (partitions[partition_id].is_locked) {
-			_SEC_HW_ERROR("%s: Error: can't delete a locked partition\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-
-		partitions[partition_id].is_created = false;
-		/* wipe the create and lock files of the partition */
-		f_open(&filep2, partitions[partition_id].create_name, FA_CREATE_ALWAYS | FA_WRITE);
-		f_close(&filep2);
-		f_open(&filep2, partitions[partition_id].lock_name, FA_CREATE_ALWAYS | FA_WRITE);
-		f_close(&filep2);
-
-		STORAGE_SET_ONE_RET(0)
-	} else if (buf[0] == STORAGE_OP_SET_CONFIG_KEY) {
-		if (is_config_locked) {
-			_SEC_HW_ERROR("%s: Error: config is locked (set config key op)\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-
-		STORAGE_GET_ZERO_ARGS_DATA
-		if (data_size != STORAGE_KEY_SIZE) {
-			_SEC_HW_ERROR("%s: Error: incorrect config key size\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-
-		for (int i = 0; i < STORAGE_KEY_SIZE; i++)
-			config_key[i] = data[i];
-
-		STORAGE_SET_ONE_RET(0)
-
-	} else if (buf[0] == STORAGE_OP_UNLOCK_CONFIG) {
-		if (!is_config_locked) {
-			STORAGE_SET_ONE_RET(0)
-			return;
-		}
-
-		STORAGE_GET_ZERO_ARGS_DATA
-		if (data_size != STORAGE_KEY_SIZE) {
-			_SEC_HW_ERROR("%s: Error: incorrect key size (sent for unlocking config)\n", __func__);
-			STORAGE_SET_ONE_RET(ERR_INVALID)
-			return;
-		}
-
-		for (int i = 0; i < STORAGE_KEY_SIZE; i++) {
-			if (config_key[i] != data[i]) {
-				STORAGE_SET_ONE_RET(ERR_INVALID);
-				return;
-			}
-		}
-
-		is_config_locked = false;
-
-		STORAGE_SET_ONE_RET(0)
-	} else if (buf[0] == STORAGE_OP_LOCK_CONFIG) {
-		is_config_locked = true;
-		STORAGE_SET_ONE_RET(0)
-	} else {
-		STORAGE_SET_ONE_RET(ERR_INVALID)
-		return;
-	}
+	XMbox_WriteBlocking(Mbox_regs[queue_id], (u32*) buf, MAILBOX_QUEUE_MSG_SIZE_LARGE);
 }
 
 void mailbox_change_queue_access_bottom_half(uint8_t queue_id)
@@ -775,48 +136,6 @@ static void handle_mailbox_interrupts(void* callback_ref)
 	_SEC_HW_DEBUG("interrupt cleared");
 }
 
-/* duplicate as in mailbox_os.c */
-_Bool handle_partial_message(uint8_t *message_buffer, uint8_t *queue_id, u32 bytes_read) 
-{
-	/* Same handling logic as in semaphore.c
-	 * If the message is incomplete due to sync issues, try to collect
-	 * the rest of the message in the next read.
-	 */
-	_SEC_HW_DEBUG1("queue %d read only %d bytes, should be %d bytes",
-		*queue_id, bytes_read, MAILBOX_QUEUE_MSG_SIZE);
-	if (!sketch_buffer[*queue_id]) {
-		_SEC_HW_DEBUG1("new sktech_buffer", *queue_id);
-		sketch_buffer_offset[*queue_id] = bytes_read;
-		sketch_buffer[*queue_id] = (uint8_t*) calloc(MAILBOX_QUEUE_MSG_SIZE, sizeof(uint8_t));
-		memcpy(sketch_buffer[*queue_id], message_buffer, bytes_read);
-		*queue_id = 0;
-		return FALSE;
-	} else {
-		/* There is already a incomplete message on the sketch_buffer */
-		if (bytes_read + sketch_buffer_offset[*queue_id] > MAILBOX_QUEUE_MSG_SIZE) {
-			_SEC_HW_ERROR("mailbox corrupted: buffer overflow");
-			_SEC_HW_ASSERT_NON_VOID(FALSE)
-		}
-
-		memcpy(sketch_buffer[*queue_id] + sketch_buffer_offset[*queue_id],
-				message_buffer, bytes_read);
-		if (bytes_read + sketch_buffer_offset[*queue_id] == MAILBOX_QUEUE_MSG_SIZE) {
-			/* This read completes the message */
-			_SEC_HW_DEBUG1("complete sketch_buffer");
-			memcpy(message_buffer, sketch_buffer[*queue_id], MAILBOX_QUEUE_MSG_SIZE);
-			free(sketch_buffer[*queue_id]);
-			sketch_buffer[*queue_id] = NULL;
-			return TRUE;
-		} else {
-			/* The message is still incomplete after this read */
-			_SEC_HW_DEBUG1("partially full sketch_buffer");
-			*queue_id = 0;
-			return FALSE;
-		}
-
-	}
-}
-
 void storage_event_loop(void)
 {
 	u32 bytes_read;
@@ -825,17 +144,7 @@ void storage_event_loop(void)
 	while(1) {
 		memset(buf, 0x0, MAILBOX_QUEUE_MSG_SIZE);
 		sem_wait(&interrupts[Q_STORAGE_CMD_IN]);
-#ifdef HW_MAILBOX_BLOCKING
 		XMbox_ReadBlocking(&Mbox_storage_cmd_in, (u32*) buf, MAILBOX_QUEUE_MSG_SIZE);
-#else
-		queue_id = Q_STORAGE_CMD_IN;
-		XMbox_Read(&Mbox_storage_cmd_in,
-			(u32*) buf,
-			MAILBOX_QUEUE_MSG_SIZE,
-			&bytes_read);
-		if (bytes_read != MAILBOX_QUEUE_MSG_SIZE &&
-			!handle_partial_message(buf, &queue_id, bytes_read))
-#endif
 		process_request(buf);
 		XMbox_WriteBlocking(&Mbox_storage_cmd_out, (u32*) buf, MAILBOX_QUEUE_MSG_SIZE);
 	}
@@ -1038,11 +347,8 @@ int init_storage(void)
 	sem_init(&interrupts[Q_STORAGE_CMD_IN], 0, 0);
 	sem_init(&interrupts[Q_STORAGE_CMD_OUT], 0, MAILBOX_QUEUE_SIZE);
 
+	initialize_ramfs();
 	initialize_storage_space();
-
-	/* set an initial config key */
-	for (int i = 0; i < STORAGE_KEY_SIZE; i++)
-		config_key[i] = i;
 
 	return XST_SUCCESS;
 }
