+++ arch/umode/include/arch/mailbox_storage.h
@@ -4,6 +4,9 @@
 void storage_event_loop(void);
 int init_storage(void);
 void close_storage(void);
+void send_response(uint8_t *buf, uint8_t queue_id);
+void read_data_from_queue(uint8_t *buf, uint8_t queue_id);
+void write_data_to_queue(uint8_t *buf, uint8_t queue_id);
 
 #endif /* __UMODE_MAILBOX_STORAGE_H */
